message : file dashboard statistics add

+++ src/main/java/com/GASB/file/model/dto/response/dashboard/FileDashboardDto.java
@@ -12,15 +12,15 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class FileDashboardDto {
-    private int total_count;
-    private double total_volume;
+    private long total_count;
+    private long total_volume;
     private int total_dlp;
     private int total_malware;
     private List<TotalTypeDto> total_type;
     private List<StatisticsDto> statistics;
 
     @Builder
-    public FileDashboardDto(int total_count, double total_volume, int total_dlp, int total_malware, List<TotalTypeDto> total_type, List<StatisticsDto> statistics){
+    public FileDashboardDto(long total_count, long total_volume, int total_dlp, int total_malware, List<TotalTypeDto> total_type, List<StatisticsDto> statistics){
         this.total_count = total_count;
         this.total_volume = total_volume;
         this.total_dlp = total_dlp;
